Bugfix: Changed `name` and `source` of `dummyData.js`
The dummyData name and source were not matched with the endpoints.
This, when the cards were changed to suit the endpoints the titles and
recipe source were left blank.

To fix this bug, I simply changed
`name` ==> `recipe_name`
`source` ==> `recipe_source`
This was just done to make sure you have a proper viewing when
you are using dummyData to test your work.